v1.50 - added collision, error logger - use ErrorLog.log(String)

+++ Hidden/src/com/cpjd/hidden/toolbox/Console.java
@@ -23,15 +23,19 @@ public class Console {
 	private boolean open;
 	private SmartField field;
 	
-	private String lastCommand = "";
+	private String lastCommand;
 	
 	private ArrayList<String> output;
+	
+	private ErrorLog errorLog;
+	private boolean errorLogOpen = false;
 	
 	private final String[] HELP = {
 		"Command, Description, Usage",
 		"stop - force close the program",
 		"reload - reload the current gamestate",
-		"clear - remove all messages from the console"
+		"clear - remove all messages from the console",
+		"log - toggles visibility of error log"
 	};
 	
 	public Console(GameStateManager gsm) {
@@ -43,6 +47,8 @@ public class Console {
 		field.setBlinkSpeed(40);
 		
 		output = new ArrayList<String>();
+		
+		errorLog = new ErrorLog();
 	}
 	
 	public void update() {
@@ -87,12 +93,18 @@ public class Console {
 			}
 			return;
 		}
+		else if (tokens[0].toLowerCase().equals("log")){
+			errorLogOpen = !errorLogOpen;
+			output.add("Error log toggled.");
+			return;
+		}
 		else{
 			output.add("Unrecognized command. Type help for list of commands");
 		}
 	}
 	
 	public void draw(Graphics2D g) {
+		
 		if(open) {
 			g.setColor(Color.WHITE);
 			g.fillRect(0, 0, 452, Layout.aligny(20));
@@ -101,12 +113,15 @@ public class Console {
 			
 			// Draw items
 			g.setColor(Color.BLACK);
+			g.setFont(new Font("Arial", Font.PLAIN, 15));
 			for(int i = output.size() - 1, j = 0; i >= 0; i--, j++) {
 				g.drawString(output.get(i), 5, Layout.aligny(19) - (j * 20));
 			}
 			
-			
-			
+		}
+		
+		if(errorLogOpen){
+			errorLog.draw(g);
 		}
 		
 		if(!GamePanel.DEBUG) return;
@@ -121,6 +136,12 @@ public class Console {
 	public boolean keyPressed(int k) {
 		if(k == KeyEvent.VK_BACK_QUOTE) open = !open;
 		if(k == KeyEvent.VK_UP){
+			
+			if(lastCommand == null){
+				output.add("No previous command");
+				return true;
+			}
+			
 			for(int i = 0; i < 50; i++) field.delete();
 			
 			lastCommand.toUpperCase();
